LayoutTests/imported/w3c/web-platform-tests/workers/constructors/Worker/AbstractWorker.onerror-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token ')'

PASS AbstractWorker.onerror
